Remove types/ directory
