+++ components/articles/index.ts
@@ -6,7 +6,7 @@ import {
   updateReformeByName,
 } from "../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../types";
+import { RootState } from "../../redux/reducers";
 import ArticlesComponent from "./articles-component";
 
 const mapStateToProps = ({ parameters }: RootState) => parameters;
