+++ components/articles/quotient-familial/regles-generales/QfTable.tsx
@@ -8,9 +8,9 @@ import { connect, ConnectedProps } from "react-redux";
 
 import { addNewLineInParameterArray, removeLastLineInParameterArray } from "../../../../redux/actions/parameters";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../../../types";
+import { RootState } from "../../../../redux/reducers";
 // eslint-disable-next-line no-unused-vars
-import { ParametersState } from "../../../../types/parameters";
+import { ParametersState } from "../../../../redux/reducers/parameters";
 import { StateParameter } from "../../../articles-inputs/parameter";
 import { Button } from "../../buttons";
 import styles from "./QfTable.module.scss";
