+++ components/dotations/results/commune-type/CommuneType.tsx
@@ -1,13 +1,20 @@
 import { Fragment, PureComponent } from "react";
 
-// eslint-disable-next-line no-unused-vars
-import { Commune } from "../../../../types/parameters";
 import { Card } from "../../../card";
 import styles from "./CommuneType.module.scss";
 import { DotationParHab } from "./dotation-par-hab";
 import { Eligibilite } from "./eligibilite";
 import { HabitantLabel } from "./habitant-label";
 import { PotentielFinancier } from "./potentiel-financier";
+
+interface Commune {
+  name: string;
+  departement: string;
+  habitants: number;
+  potentielFinancier: number;
+  eligible: boolean;
+  dotationParHab: number;
+}
 
 export class CommuneType extends PureComponent<Commune> {
   render() {
