+++ components/simulation-page/SimulationPage.tsx
@@ -6,7 +6,7 @@ import { PureComponent } from "react";
 import SwipeableViews from "react-swipeable-views";
 
 // eslint-disable-next-line no-unused-vars
-import { ParametersState } from "../../types/parameters";
+import { ParametersState } from "../../redux/reducers/parameters";
 import AppHeader from "../app-header";
 import { ArticleHeader } from "./article-header";
 import { SimulationMenuBar } from "./simulation-menu";
