+++ redux/actions/parameters/remove-last-line-in-parameter-array.ts
@@ -1,6 +1,7 @@
-// eslint-disable-next-line no-unused-vars
-import { RemoveLastLineInParameterArrayAction } from "../../../types/parameters";
-
+export interface RemoveLastLineInParameterArrayAction {
+  type: "REMOVE_LAST_LINE_IN_PARAMETER_ARRAY",
+  path: string;
+}
 export function removeLastLineInParameterArray(path: string): RemoveLastLineInParameterArrayAction {
   return {
     path,
