+++ redux/actions/parameters/reset-amendement-to-plf.ts
@@ -1,5 +1,10 @@
 // eslint-disable-next-line no-unused-vars
-import { ResetAmendementToPlfAction } from "../../../types/parameters";
+import { ParametersState } from "../../reducers/parameters";
+
+export interface ResetAmendementToPlfAction {
+  type: "RESET_AMENDEMENT_TO_PLF",
+  topic: keyof ParametersState,
+}
 
 export function resetAmendementToPlf(topic: ResetAmendementToPlfAction["topic"]): ResetAmendementToPlfAction {
   return {
