+++ redux/actions/parameters/update-parameter.ts
@@ -1,5 +1,8 @@
-// eslint-disable-next-line no-unused-vars
-import { UpdateParameterAction } from "../../../types/parameters";
+export interface UpdateParameterAction {
+  type: "UPDATE_PARAMETER";
+  path: string;
+  value: any;
+}
 
 export function updateParameter(path: string, value: any): UpdateParameterAction {
   return {
