+++ redux/reducers/parameters/index.ts
@@ -4,6 +4,8 @@ import { amendement } from "./amendement";
 import { base } from "./base";
 import { plf } from "./plf";
 
+export * from "./parameters-state";
+
 export const parameters = combineReducers({
   amendement,
   base,
