--- types/parameters/actions.ts
@@ -1,35 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import { ParametersState } from "./parameters-state";
-
-export interface UpdateParameterAction {
-  type: "UPDATE_PARAMETER";
-  path: string;
-  value: any;
-}
-
-export interface AddNewLineInParameterArrayAction {
-  type: "ADD_NEW_LINE_IN_PARAMETER_ARRAY",
-  path: string;
-}
-
-export interface RemoveLastLineInParameterArrayAction {
-  type: "REMOVE_LAST_LINE_IN_PARAMETER_ARRAY",
-  path: string;
-}
-
-export interface ResetAmendementToBaseAction {
-  type: "RESET_AMENDEMENT_TO_BASE",
-  topic: keyof ParametersState,
-}
-
-export interface ResetAmendementToPlfAction {
-  type: "RESET_AMENDEMENT_TO_PLF",
-  topic: keyof ParametersState,
-}
-
-export type ParameterAction =
-  UpdateParameterAction |
-  AddNewLineInParameterArrayAction |
-  RemoveLastLineInParameterArrayAction |
-  ResetAmendementToBaseAction |
-  ResetAmendementToPlfAction;
