--- types/parameters/commune.ts
@@ -1,8 +1,0 @@
-export interface Commune {
-  name: string;
-  departement: string;
-  habitants: number;
-  potentielFinancier: number;
-  eligible: boolean;
-  dotationParHab: number;
-}
