--- types/parameters/index.ts
@@ -1,3 +1,0 @@
-export * from "./actions";
-export * from "./commune";
-export * from "./parameters-state";
