--- types/root-state.ts
@@ -1,3 +1,0 @@
-import rootReducer from "../redux/reducers";
-
-export type RootState = ReturnType<typeof rootReducer>;
